requirements.txt
ultralytics
easydict==1.11
einops==0.7.0
matplotlib
mmcv_full==1.7.1
numpy
opencv_python
pandas==1.3.5
Pillow
psutil==5.8.0
pytest==7.4.3
scikit_learn==1.0.2
scipy==1.7.3
setuptools==59.5.0
timm==0.9.8
tqdm==4.65.2
ttach==0.0.3
visdom==0.2.4
yacs==0.1.8
tensorboard
torchsummary
onnx
onnxruntime-gpu
